skbuff: Add bit indicating the fragment neeeds special dma handlng.
zctap pages may refer to system pages which are mapped into user
space, or dummy pages which are not mapped anywhere.

The first type requires DMA sync between the cpu and device.
The second type uses a lookup table to get the dma address.

Use a single bit to indicate special handling is required,

Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>