net: Add a stream_memory_rdband() callback
Protocols have a stream_memory_read() callback which set
EPOLLRDNORM if something is readable on the socket.

When sending on on ZC socket, often a completion notification
is generated when the send is complete.  Add an optional
stream_memory_rdband() callback which generates EPOLLRDBAND,
this is used to indicate readability on a secondary receive
queue.

Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>