zctap/tcp: allow per-socket tcp option padding
Not for upstream.

Introduces a TP socket option TCP_HDR_MINSIZE, which allows padding
out tcp headers with NOPs to a fixed size.  This is intended for use
on the TCP sender side, so a TCP receiver gets fixed size headers.

This is a workaround hack for network cards which cannot do true
header splitting, and fake it with byte counting instead.  This hack
is not needed if the backing store is accessible to the CPU, as the
receiver then can perform pskb_pull(), moving data from overly large
headers into the linear area.

Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>